chore(Fox-H.pdf): update formatting of bibliography entries
feat(GenBib.sh): add script to generate bibliography for Fox-H.tex

The GenBib.sh script is added to generate the bibliography for the Fox-H.tex document. The script removes any existing auxiliary files and the Fox-H_biber.bib file. It then runs lualatex with the --shell-escape option to compile the Fox-H.tex document. After that, it uses biber to generate the .bbl file and resolve any references in the Fox-H.bcf file. The script then runs lualatex again to update the document with the generated bibliography. Finally, it uses bibtex-tidy to format the Fox-H_biber.bib file and removes all auxiliary files.